Srss model

+++ prlab/gutils.py
@@ -28,8 +28,9 @@ def make_check_point_folder(config={}, cp_base=None, cp_name=None):
     csvLog = cp_path / "loger"
 
     # write configure to easy track later, could not write JSON because some object inside config
+    cp_path.mkdir(exist_ok=True)
     txtwriter = cp_path / 'configure.txt'
-    txtwriter.write_text(str(config))
+    txtwriter.write_text(str(config)) if not txtwriter.is_file() else None
 
     return cp_path, best_name, csvLog
